+++ lib/ayeaye/model_collection.py
@@ -1,304 +1,250 @@
 from collections import defaultdict
-import copy
 from dataclasses import dataclass
 from inspect import isclass
 
-from ayeaye.connectors.base import AccessMode
-from ayeaye.connectors.multi_connector import MultiConnector
 from ayeaye.model import Model
-from ayeaye.pinnate import Pinnate
 
 
 class ModelCollection:
-    def __init__(self, models):
+    def __init__(self, models, class_level_eval=False):
         """
-        Build graphs from a group of subclasses of :class:`ayeaye.Models`
+        Build graphs from a group of subclasses of :class:`ayeaye.Models`.
 
-        @param models (class, list, set or callable)
-            All of which are :class:`ayeaye.Models`. They aren't instances of the class but the
-            class itself. A *list* is an explicit ordering that the models should be run in. Models
-            in *set* will be examined to see dataset dependencies and these will be used to
-            determine the model run order.
+        Use the graph of models to
+        * find data provenance
+        * interactively determine which models have their data dependences satisfied and are
+          ready to run.
+
+        @param models (set of subclases of :class:`ayeaye.Models`).
+            These aren't instances of the class but the class itself.
+
+        @param class_level_eval: (bool) When True, just use the class declarations to determine the
+            relationships between classes. This avoids prepare dataset connectors which often requires
+            the connector_resolver. See :class:`ayeaye.connect_resolve.ConnectorResolver` for more.
         """
-        super().__init__()
 
         invalid_construction_msg = (
-            "models must be a class, list, set or callable. All of which "
-            "result in one or more :class:`ayeaye.Models` classes (not "
-            "instances)."
+            "models must be a class, list, set or callable. All of which result in one or more "
+            ":class:`ayeaye.Models` classes (not instances)."
         )
 
-        # validate and prepare
-        if isclass(models) and issubclass(models, Model):
-            models = [
-                models,
-            ]
-
-        if callable(models):
-            # TODO
-            raise NotImplementedError("TODO")
-
-        elif isinstance(models, (list, set)):
-            if not all([isclass(m) and issubclass(m, Model) for m in models]):
-                raise ValueError(invalid_construction_msg)
-            self.models = models
-
-        else:
+        if not isinstance(models, set):
             raise ValueError(invalid_construction_msg)
 
-    def __len__(self):
-        return len(self.models)
+        if not all([isclass(m) and issubclass(m, Model) for m in models]):
+            raise ValueError(invalid_construction_msg)
 
-    def __iter__(self):
+        self.models = models
+        self.class_level_eval = class_level_eval
+
+        # see :meth:`ready_to_run`
+        self._completed_models = set()
+
+        # lazy, see :meth:`model_dependencies`
+        self._model_dependencies = None
+
+    @property
+    def model_dependencies(self):
+
+        if self._model_dependencies is None:
+            self._model_dependencies = self._resolve_model_dependencies()
+
+            # this raises an exception if there is a cycle
+            assert self._valid_model_dependencies(self._model_dependencies)
+
+        return self._model_dependencies
+
+    def ready_to_run(self, just_completed=None):
         """
-        yield model classes. No ordering when models are a set but does honour order when models
-        is a list
+        Note - this method doesn't take into account models that are currently running.
+
+        @param just_completed: (set of model classes) signal that these models have been completed.
+            The return to this method will take these into account.
+
+        @return: set of `self.models` that have had all dependencies satisfied.
         """
-        yield from self.models
+
+        if just_completed:
+            for model_cls in just_completed:
+                if model_cls not in self.models:
+                    raise ValueError(f"Model {model_cls.__name__} isn't known to ModelCollection")
+
+                self._completed_models.add(model_cls)
+
+        ready_models = set()
+        for model_cls, depends_on in self.model_dependencies:
+
+            if model_cls in self._completed_models:
+                # model has already been run
+                continue
+
+            if depends_on.issubset(self._completed_models):
+                ready_models.add(model_cls)
+
+        return ready_models
 
     def _base_graph(self):
+
+        if self.class_level_eval:
+            raise NotImplementedError("TODO _base_graph_no_resolve")
+        else:
+            return self._base_graph_with_resolve()
+
+    def _base_graph_with_resolve(self):
         """Find all the datasets in all the models and classify datasets as 'sources' (READ access)
         and 'targets' (WRITE access) or READWRITE for both.
 
-        @return: (targets, sources, nodes) (set, set, dictionary)
+        This version of _base_graph needs the connect_resolve in order to evaluate engine_urls.
 
-            node - 'datasets' are a subclasses of :class:`Connect` or :class:`DataConnector`
+        @return: (targets, sources) (dict, dict) - data_flow_key -> set of tuples
+                (ModelCls, dataconnector attrib name) for both
 
-            all_targets - one or more models writes to these
-            all_sources - one or more models read from these
-            nodes - key is model class, value is Pinnate with .model_cls, .model_name, .targets, .sources
+            targets - one or more models writes to these
+            sources - one or more models read from these
+
         """
-        all_targets = set()
-        all_sources = set()
-        nodes = {}
+        # data_flow_key -> (ModelCls, dataconnector attrib name)
+        targets = defaultdict(set)
+        sources = defaultdict(set)
+
+        # ensure resolution to a name string doesn't contain clashes
+        check_names = set()
+
         for model_cls in self.models:
             # TODO find ModelConnectors and recurse into those
             model_name = model_cls.__name__
-            if model_name in nodes:
+            if model_name in check_names:
                 raise ValueError(f"Duplicate node found: {model_name}")
+            check_names.add(model_name)
 
-            node = Pinnate(
-                {
-                    "model_cls": model_cls,
-                    "model_name": model_name,
-                    "targets": set(),
-                    "sources": set(),
-                }
-            )
+            # Not instantiating the model `model_obj = model_cls()` because it's not needed
+            # in order to find data dependencies.
+            for class_attrib_label, connect in model_cls.connects().items():
 
-            # as instantiated model
-            # model_obj = model_cls()
+                # this requires connect_resolve
+                data_connector = connect._prepare_connection()
 
-            for class_attrib_label, ds_connector in model_cls.connects().items():
-                # for class_attrib_label, ds_connector in model_obj.datasets().items():
+                data_flow = data_connector.data_flow()
 
-                # if isinstance(ds_connector, Connect):
-                #
-                #     if isinstance(ds_connector, list):
-                #         # experiment with Connector prior to instantiation as MultiConnector
-                #         m_connectors = [Connect()]
+                for data_flow_key in data_flow.inputs:
+                    sources[data_flow_key].add((model_cls, class_attrib_label))
 
-                if isinstance(ds_connector, MultiConnector):
-                    m_connectors = ds_connector.data
-                else:
-                    m_connectors = [ds_connector]
+                for data_flow_key in data_flow.outputs:
+                    targets[data_flow_key].add((model_cls, class_attrib_label))
 
-                for dataset_connector in m_connectors:
-                    dataset_container = ModelDataset(
-                        model_attrib_label=class_attrib_label, connector=dataset_connector
-                    )
+        return targets, sources
 
-                    if dataset_connector.access in [AccessMode.READ, AccessMode.READWRITE]:
-                        node.sources.add(dataset_container)
-                        all_sources.add(dataset_container)
-
-                    if dataset_connector.access in [AccessMode.WRITE, AccessMode.WRITE]:
-                        node.targets.add(dataset_container)
-                        all_targets.add(dataset_container)
-
-            nodes[model_cls] = node
-
-        return all_targets, all_sources, nodes
-
-    def _resolve_run_order(self):
+    def _resolve_model_dependencies(self):
         """
         Use the dataset connections in each model to determine the dependencies between models and
         therefore the order to run them in.
 
-        @return: (Pinnate) with attributes:
-                    leaf_sources (set of datasets) - read but not written to
-                    leaf_targets (set of datasets) - written to but not read - end goal of model
-                    run_order (list of sets of nodes) - All models in a set can be run in parallel.
-                            Each set must be complete before the next set is run.
+        @return: list of tuples (Model, {Models dependencies}
+                dependencies are those models which must be completed before `Model` can run.
 
-                    'datasets' as subclasses of :class:`Connect`
-                    `nodes` are type :class:`Pinnate` with attributes model_cls, model_name,
-                        targets and sources
         """
-        # This algorithm is a bit overly simplistic and sub-optimal. The return is a list of sets
-        # with all models in set needing to be run. But the set is actually the models that were
-        # ready to run when the prior models are complete. So there could be subsequent models that
-        # are only waiting on a subset of each set.
-        all_targets, all_sources, nodes = self._base_graph()
+        targets, sources = self._base_graph()
 
-        leaf_sources = all_sources - all_targets
-        leaf_targets = all_targets - all_sources
+        r = []
+        for model_cls in self.models:
 
-        completed = copy.copy(leaf_sources)
-        run_order = []
-        while len(nodes) > 0:
-            loop_ready = set()
-            for node in nodes.values():
-                if node.sources.issubset(completed):
-                    loop_ready.add(node)
+            # no attempt at efficiency
 
-            if len(loop_ready) == 0:
-                msg = "The set of models can't be built into a single acyclic graph."
-                raise ValueError(msg)
+            data_deps = set()
+            for data_flow_key, deps in sources.items():
+                for dep_model_cls, _class_attrib_label in deps:
+                    if model_cls == dep_model_cls:
+                        data_deps.add(data_flow_key)
 
-            # nodes in loop_ready have been visited add their edges to the completed map
-            run_order.append(loop_ready)
-            for ready_node in loop_ready:
-                node = nodes[ready_node.model_cls]
-                completed.update(node.targets)
-                del nodes[ready_node.model_cls]
+            model_deps = set()
+            for data_flow_key, target_deps in targets.items():
 
-        p = Pinnate(
-            {"leaf_sources": leaf_sources, "leaf_targets": leaf_targets, "run_order": run_order}
-        )
-        return p
+                if data_flow_key not in data_deps:
+                    continue
 
-    def run_order(self):
+                for dep_model_cls, _class_attrib_label in target_deps:
+                    model_deps.add(dep_model_cls)
+
+            # READWRITE results in a model being dependent on itself so remove that
+            model_deps.discard(model_cls)
+
+            # print(model_cls.__name__, data_deps, [m.__name__ for m in model_deps])
+            r.append((model_cls, model_deps))
+
+        return r
+
+    def _valid_model_dependencies(self, model_deps):
         """
-        @return: (list) items in the list are either-
-                    (i) sets of subclasses of :class:`ayeaye.Model`
-                    OR
-                    (ii) lists of sets as per (i)
+        raise ValueError if there is any funny cyclic business going on.
 
-        All models in a set can be run in parallel. Each set must be complete before the next set
-        in the list is run.
+        @param model_deps: from :meth:`_resolve_model_dependencies`
         """
-        if isinstance(self.models, list):
-            # models in 'list' mode are wrapped as single item sets because all items in a set must
-            # complete before next item in a list is run.
-            return [{m} for m in self]
+        deps_lookup = {model_cls: model_d for model_cls, model_d in model_deps}
 
-        elif isinstance(self.models, set):
-            # _resolve_run_order returns leaf nodes and the run order is built using Pinnate
-            # instances as nodes. Remove all this with to return a simple run order.
-            resolved_order = self._resolve_run_order()
+        def build_path(deps_lookup, m, ancestors):
+            "Find all paths. Not really needed, just the detection of cycles."
 
-            def simplify_nodes(n):
-                if isinstance(n, Pinnate):
-                    return n.model_cls
-                if isinstance(n, (set, list)):
-                    rx = [simplify_nodes(nx) for nx in n]
-                    if isinstance(n, set):
-                        return set(rx)
-                    return rx
+            new_nodes = []
+            for m2 in deps_lookup.get(m, []):
+                if m2 in ancestors:
+                    raise ValueError("Cycle detected")
+                new_nodes.append(m2)
 
-            return [simplify_nodes(r) for r in resolved_order.run_order]
+            if len(new_nodes) == 0:
+                yield ancestors + [m]
+            else:
+                for mx in new_nodes:
+                    yield from build_path(deps_lookup, mx, ancestors + [m])
 
-        else:
-            raise ValueError("Unknown models container.")
+        for model_cls, _model_d in model_deps:
 
-    def dataset_provenance(self):
+            _all_paths = [p for p in build_path(deps_lookup, model_cls, [])]
+
+            # for p in _all_paths:
+            #     debug_path = "-".join([px.__name__ for px in p])
+            #     print(debug_path)
+
+        return True
+
+    def as_edges(self):
         """
-        self.models is a set of :class:`ayeaye.Model`s. These might all be interconnected or they
-        may form multiple graphs.
+        Layout model dependencies as graph edges.
 
-        @return: set (graphs) of sets (edges)
-            graphs are a set of edges
-            edges are :class:`ModelGraphEdge` objects.
+        @param model_deps: from :meth:`_resolve_model_dependencies`
+        @return: set of :class:`ModelGraphEdge`
         """
-        all_targets, all_sources, nodes = self._base_graph()
-        leaf_sources = all_sources - all_targets
-        leaf_targets = all_targets - all_sources
-
-        # lookup to sources for each dataset
-        dataset_sources = defaultdict(list)
-        for node in nodes.values():
-            for dataset_container in node.sources:
-                dataset_sources[dataset_container].append(node.model_cls)
-
-        def dataset_source(target_dataset_container):
-            "generator yielding model classes"
-            for source_model_cls in dataset_sources[target_dataset_container]:
-                for source_dataset_container in nodes[source_model_cls].sources:
-                    if source_dataset_container == target_dataset_container:
-                        yield source_model_cls, source_dataset_container
+        # TODO duplicate call to _base_graph
+        # targets, sources = self._base_graph()
+        model_deps = self._resolve_model_dependencies()
+        #
+        # # Model_cls -> dict( sources : list, targets : list)
+        # # sources and targets are list of data_flow_key (str)
+        # datasets_lookup = {}
+        # for label, rels in [("targets", targets), ("sources", sources)]:
+        #     for data_flow_key, model_att in rels.items():
+        #         for model_cls, _class_attrib_label in model_att:
+        #             if model_cls not in datasets_lookup:
+        #                 datasets_lookup[model_cls] = defaultdict(list)
+        #
+        #             datasets_lookup[model_cls][label].append(data_flow_key)
 
         edge_set = set()
-        for node in nodes.values():
-            # print(node.model_cls.__name__, len(node.targets))
+        for model_cls, model_depends_on in model_deps:
+            for model_d in model_depends_on:
+                edge = ModelGraphEdge(
+                    model_a=model_d,
+                    model_b=model_cls,
+                    dataset_label="",  # data_flow_key here would be ace
+                )
+                edge_set.add(edge)
 
-            for dataset_container_a in node.sources:
-                if dataset_container_a in leaf_sources:
-                    dataset_label = dataset_container_a.model_attrib_label
-                    mge = ModelGraphEdge(
-                        model_a=None, model_b=node.model_cls, dataset_label=dataset_label
-                    )
-                    edge_set.add(mge)
-
-            for dataset_container_a in node.targets:
-                if dataset_container_a in leaf_targets:
-                    dataset_label = dataset_container_a.model_attrib_label
-                    mge = ModelGraphEdge(
-                        model_a=node.model_cls, model_b=None, dataset_label=dataset_label
-                    )
-                    edge_set.add(mge)
-
-                for model_b, dataset_container_b in dataset_source(dataset_container_a):
-                    # print(node.model_cls, dataset_container_a, model_b, dataset_container_b)
-
-                    dataset_label = dataset_container_a.model_attrib_label
-                    # models might each use different attrib names
-                    if (
-                        dataset_container_b is not None
-                        and dataset_label != dataset_container_b.model_attrib_label
-                    ):
-                        dataset_label += " / " + dataset_container_b.model_attrib_label
-
-                    mge = ModelGraphEdge(
-                        model_a=node.model_cls, model_b=model_b, dataset_label=dataset_label
-                    )
-                    edge_set.add(mge)
-
-        # TODO - are sub graphs needed?
-        graph_set = list()
-        graph_set.append(list(edge_set))
-        return graph_set
-
-
-@dataclass
-class ModelDataset:
-    """
-    Container for attributes associated with a dataset that belongs to a model.
-
-    A dataset could be a :class:`ayeaye.Connect` or it could have been resolved to a subclass of
-    :class:`DataConnector`. Hold these (todo) and other associated info about the dataset.
-    """
-
-    model_attrib_label: str  # name of class variable in parent model
-    connector: object  #  ayeaye.Connect
-
-    def __hash__(self):
-        """
-        super critical to graphs being able to build is equating two datasets are the same thing
-        without the name (i.e. model_attrib_label) mattering.
-        """
-        return self.connector.__hash__()
-
-    def __eq__(self, other):
-        if isinstance(other, type(self)):
-            return self.connector == other.connector
-        return False
+        return edge_set
 
 
 @dataclass
 class ModelGraphEdge:
     """
-    A representation of the :class:`ModelGraph` is a list of edges.
+    A representation of the datasets and models as is a list of edges.
 
     Two :class:`ayeaye.Model`s and a 'dataset' (:class:`Connect` or `DataConnector`) is an edge.
     This was an arbitrary decision, it could have been two datasets and a model.
@@ -338,21 +284,16 @@ class VisualiseModels:
         # no idea if leaves are the same dataset or not
         leaf_label = _leaf_label()
 
-        graphset = self.model_collection.dataset_provenance()
-        if len(graphset) == 0:
+        edgeset = self.model_collection.as_edges()
+        if len(edgeset) == 0:
             return ""
 
-        # if len(graphset) > 1:
-        #     raise NotImplementedError("Multiple graphs within one collection of models is not yet implemented!")
-
         out = ["graph LR"]
-        for graph in graphset:
-            # graphs don't need to be separate for Mermaid
-            for edge in graph:
-                model_a = edge.model_a.__name__ if edge.model_a is not None else next(leaf_label)
-                model_b = edge.model_b.__name__ if edge.model_b is not None else next(leaf_label)
+        for edge in edgeset:
+            model_a = edge.model_a.__name__ if edge.model_a is not None else next(leaf_label)
+            model_b = edge.model_b.__name__ if edge.model_b is not None else next(leaf_label)
 
-                edge_fmt = f"{model_a}-->|{edge.dataset_label}| {model_b}"
-                out.append(edge_fmt)
+            edge_fmt = f"{model_a}-->|{edge.dataset_label}| {model_b}"
+            out.append(edge_fmt)
 
         return "\n".join(out)